获取单个收货地址bug

+++ service/src/main/java/com/zwl/service/UserReceivingAddressService.java
@@ -20,7 +20,7 @@ public interface UserReceivingAddressService {
 
     void updateIsDefaultByUserId(String userId, String merchantId);
 
-    UserReceivingAddress getOneById(Long id);
+    UserReceivingAddress getOneById(Integer id);
 
     void deleteById(Integer id);
 }
